[#1853] - Separa el puerto del adaptador Sanity en el modulo repository (DIP)
El doble in-memory importaba projectSingleSection del archivo del adaptador
concreto (literary-work.repository.ts alojaba la interfaz junto a Sanity-
LiteraryWorkRepository) — un adaptador dependiendo de otro, sintoma de que el
puerto no tenia hogar propio (viola DIP: la interfaz vive con quien la usa, no
con quien la implementa).

Se separa el trio en tres archivos del modulo:
- literary-work.repository.ts: el PUERTO (interfaz LiteraryWorkRepository +
  projectSingleSection, la proyeccion dominio->dominio que es logica de contrato
  compartida por ambos adaptadores).
- literary-work.repository.sanity.ts: el ADAPTADOR Sanity (SanityLiteraryWork-
  Repository + su ACL privada + materializacion). Spec renombrado a .sanity.spec.ts.
- literary-work.repository.mock.ts: el DOBLE (InMemoryLiteraryWorkRepository), sin
  cambios: ya importaba del puerto, que ahora deja de contener al adaptador Sanity.

Ambos adaptadores dependen del puerto, ninguno del otro (ADP): el acople queda
estructuralmente imposible, no solo corregido. Fija el patron de nombres
<dominio>.repository.ts / .repository.sanity.ts / .repository.mock.ts para los
proximos modulos backend. Documentado en LITERARY_WORK_DESIGN.md §6.
