+++ docs/LITERARY_WORK_DESIGN.md
@@ -238,7 +238,7 @@ export function applyReadingTimeMaterialization(
 // campos faltantes (`isEmpty`) no hace round-trip de red.
 ```
 
-> **Estado de implementación:** el schema (campos opcionales, arrancan vacíos), esta unidad de cómputo + persistencia (#1986), y el **repository** que la cablea ya existen en `develop`/en review. `SanityLiteraryWorkRepository` (`src/api/modules/literary-work/literary-work.repository.ts`) es dueño de la traducción raw Sanity → dominio (la ACL vive **dentro del repository**, como métodos privados, no como un mapper separado en `_utils/`) y cablea la materialización self-healing invocando `buildReadingTimeMaterialization`/`applyReadingTimeMaterialization` en la primera lectura de cada obra. Es el **Slice 1a**, en review en **PR #2002**, desde `develop` (el PR #1929 fue solo el draft del modelo de #1852, no la base de este incremento). Lo que queda: el **service** (`getLiteraryWorkBySlug`) y el `?section=N` eficiente que se apoya en el repository ([§7](#7-contrato-del-endpoint)).
+> **Estado de implementación:** el schema (campos opcionales, arrancan vacíos), esta unidad de cómputo + persistencia (#1986), y el **repository** que la cablea ya existen en `develop`/en review. `SanityLiteraryWorkRepository` (`src/api/modules/literary-work/literary-work.repository.sanity.ts`) es dueño de la traducción raw Sanity → dominio (la ACL vive **dentro del repository**, como métodos privados, no como un mapper separado en `_utils/`) y cablea la materialización self-healing invocando `buildReadingTimeMaterialization`/`applyReadingTimeMaterialization` en la primera lectura de cada obra. Es el **Slice 1a**, en review en **PR #2002**, desde `develop` (el PR #1929 fue solo el draft del modelo de #1852, no la base de este incremento). Lo que queda: el **service** (`getLiteraryWorkBySlug`) y el `?section=N` eficiente que se apoya en el repository ([§7](#7-contrato-del-endpoint)).
 
 ### Duración de obras recitadas/audiovisuales
 
@@ -278,6 +278,16 @@ interface LiteraryWorkRepository {
 ```
 
 **La ACL vive dentro del repository, no en `_utils/`.** A diferencia del resto de los módulos backend (ver [`sanity-acl.md`](../.claude/references/sanity-acl.md)), `SanityLiteraryWorkRepository` no delega la traducción raw → dominio a funciones separadas en `src/api/_utils/*.functions.ts`: la implementa como **métodos privados propios** (`mapWork`, `mapMetadata`, `mapSection`, `mapEpigraph`, …), reutilizando como building blocks los sub-mappers transversales ya existentes (`mapAuthor`, `mapResources`, `mapTags`, `mapMediaSources`). Es una **divergencia intencional**: el repository es la única frontera que conoce el shape de Sanity y entrega `LiteraryWork` de dominio listo — la dirección arquitectónica objetivo del backend (repository dueño de su propia ACL), no todavía adoptada por `story` ni por el resto de los módulos. `InMemoryLiteraryWorkRepository` almacena y devuelve dominio (`LiteraryWork[]`), no crudo. Implementación en review en **PR #2002** (Slice 1a).
+
+**Estructura de archivos (puerto / adaptador / doble separados).** Para que el doble no dependa del adaptador concreto —DIP: la interfaz vive con quien la _usa_, no con quien la _implementa_— el trío se reparte en tres archivos del módulo, en vez de alojar la interfaz junto a un implementador:
+
+| Archivo                              | Rol                                                                                                                                                                                                                                                                      |
+| ------------------------------------ | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------ |
+| `literary-work.repository.ts`        | **Puerto**: la interfaz `LiteraryWorkRepository` + `projectSingleSection` (proyección dominio→dominio; es lógica del _contrato_ —parte de lo que significa implementar `fetchSectionBySlug`—, no un detalle de Sanity, así que la comparten ambos adaptadores desde acá) |
+| `literary-work.repository.sanity.ts` | **Adaptador Sanity**: `SanityLiteraryWorkRepository` + su ACL privada (traducción raw→dominio) + materialización write-on-read                                                                                                                                           |
+| `literary-work.repository.mock.ts`   | **Doble**: `InMemoryLiteraryWorkRepository` (almacena dominio)                                                                                                                                                                                                           |
+
+Ambos adaptadores importan del **puerto**, nunca entre sí (ADP: ningún adaptador conoce al otro). Es el patrón de nombres (`<dominio>.repository.ts` / `.repository.sanity.ts` / `.repository.mock.ts`) a reutilizar en los próximos módulos backend que adopten "repository dueño de su ACL".
 
 Firma del módulo backend (service, consume el repository ya en dominio):
